Moved all decoders used for structural features into own package.
Updated javadoc
